requirements.txt
streamlit
pinecone-client
sentence-transformers
langchain
pypdf
transformers
torch